==> nobel_physics_prizes/notebooks/1.2-collect-physicists-raw-data.ipynb <==
# Collect Physicists Raw Data

The goal of this notebook is to collect demographic data on the list of [physicists notable for their achievements](../data/raw/physicists.txt). Wikipedia contains this semi-structured data in an *Infobox* on the top right side of the article for each physicist. However, similar data is available in a more machine readable, [JSON](https://www.json.org/) format from [DBpedia](https://wiki.dbpedia.org/about). We will need to send HTTP requests to DBpedia to get the JSON data. For an example, compare *Albert Einstein's* [Wikipedia infobox](https://en.wikipedia.org/wiki/Albert_Einstein) to his [DBPedia JSON](http://dbpedia.org/data/Albert_Einstein.json). It is important to realize, that although the data is similar, it is not identical.

The shortcomings of Wikipedia infoboxes and the advantages of DBpedia datasets are explained in section 4.3 of [DBpedia datasets](https://wiki.dbpedia.org/services-resources/datasets/dbpedia-datasets#h434-10). But basically the summary is that DBpedia data is much cleaner and better structured than Wikipedia Infoboxes as it is based on hand-generated mappings of Wikipedia infoboxes / templates to a [DBpedia ontology](https://wiki.dbpedia.org/services-resources/ontology). Consequently, we will be using DBpedia as the data source for this project.

However, DBpedia does have the disadvantage that its content is roughly 6-18 months behind updates applied to Wikipedia content. This is due to its data being generated from a [static dump of Wikipedia content](https://wiki.dbpedia.org/online-access/DBpediaLive) in a process that takes approximately 6 months. The fact that the data is not in sync with the latest Wikipedia content is not of great significance for this project as the data is edited infrequently. Also when edits are made, they tend to be only minor.

## Setting the Environment

A few initialization steps are needed to setup the environment:

- The locale needs to be set for all categories to the user’s default setting (typically specified in the LANG environment variable) to enable correct sorting of physicists names with accents.
- A bool constant `FETCH_JSON_DATA` needs to be set to decide whether to fetch the json data. Set to False so that the previously fetched data is used. In this case the results of the study are guaranteed be reproducible. Set to True so that the latest data is fetched. In this case it is possible that the results of the study will change.

import locale
    
locale.setlocale(locale.LC_ALL, '')

FETCH_JSON_DATA = False

## Constructing the URLs

To make the HTTP requests, we will need a list of URLs representing the resources (i.e the physicists). It's fairly easy to construct these URLs from the list of notable physicists. However, it's important to "quote" any physicist name in unicode since unicode characters are not allowed in URLs. OK let's create the list now.

import gzip
import os
import shutil
from collections import OrderedDict

import jsonlines
import pandas as pd

from src.data.jsonl_utils import read_jsonl
from src.data.url_utils import DBPEDIA_DATA_URL
from src.data.url_utils import fetch_json_data
from src.data.url_utils import urls_progress_bar

def construct_urls(file='../data/raw/physicists.txt'):
    """Construct DBpedia data URLs from list in file.

    Args:
        file (str): File containing a list of url filepaths
            with spaces replacing underscores.
    Returns:
        list(str): List of URLs.

    """

    with open(file, encoding='utf-8') as file:
        names = [line.rstrip('\n') for line in file]

    urls = [DBPEDIA_DATA_URL + name.replace(' ', '_') + '.json'
            for name in names]
    return urls

urls_to_fetch = construct_urls()
assert(len(urls_to_fetch) == 1069)

## Fetching the Data

Now we have the list of URLs, it's time to make the HTTP requests to acquire the data. The code is asynchronous, which dramatically helps with performance. It is important to set the `max_workers` parameter sensibly in order to crawl responsibly and not hammer the site's server. Although the site seems to be rate limited, it's still good etiquette.

jsonl_file = '../data/raw/physicists.jsonl'
if FETCH_JSON_DATA:
    json_data = fetch_json_data(urls_to_fetch, max_workers=20, timeout=30,
                                progress_bar=urls_progress_bar(len(urls_to_fetch)))
else:
    json_data = read_jsonl('../data/raw/physicists.jsonl' + '.gz')

Let's sort the data alphabetically by URL, confirm that all the data was fetched and take a look at the first JSON response.

if FETCH_JSON_DATA:
    json_data = OrderedDict(sorted(json_data.items(), key=lambda x: locale.strxfrm(x[0])))
    assert(len(json_data) == 1069)
    print(list(json_data.keys())[0])
    print(list(json_data.values())[0])
else:
    assert(len(json_data) == 1058)
    print(json_data[0])

It is clear that every request successfully received a response. However, we see that some responses came back empty from the server. Basically, although there are Wikipedia pages for these physicists, they do not have a corresponding page in DBpedia (or the page in DBpedia has a different name). Not to worry, there are only 11 and they are not so famous, so we will just exclude these "Z-listers" from the analysis.

if FETCH_JSON_DATA:
    urls_to_drop = [url for (url, data) in json_data.items() if not data]
    assert(len(urls_to_drop) == 11)
    display(urls_to_drop)

if FETCH_JSON_DATA:
    json_data = [data for data in json_data.values() if data]
    assert(len(json_data) == 1058)

## Persisting the Data

Now that we have the list of JSON responses, we would like to persist them for later analysis. We will use [Json Lines](http://jsonlines.org/) as it seems like a convenient format for storing structured data that may be processed one record at a time.

if FETCH_JSON_DATA:
    with jsonlines.open(jsonl_file, 'w') as writer:
        writer.write_all(json_data)

Let's do a quick sanity check to make sure the file contains the expected number of records.

if FETCH_JSON_DATA:
    json_lines = read_jsonl(jsonl_file)
    assert(len(json_lines) == 1058)

Finally, let's compress the file to reduce its footprint.

if FETCH_JSON_DATA:
    with open(jsonl_file, 'rb') as src, gzip.open(jsonl_file + '.gz', 'wb') as dest:
        shutil.copyfileobj(src, dest)
    os.remove(jsonl_file)
