==> nobel_physics_prizes/notebooks/2.0-process-physicists-raw-data.ipynb <==
# Process Physicists Raw Data

Here we will convert the JSON lines [physicists raw data](../data/raw/physicists.jsonl.gz) into a pandas dataframe that is more convenient to work with. We will refer to this as an intermediate format as it will not be the final format of the data that we will be working with for analysis.

The goal is to parse the JSON data in order to extract the interesting fields of information such as *name*, *birth date*, *death date*, *citizenships*, *workplaces*, *awards*, *alma mater*, *academic advisors*, *doctoral students*, *DBpedia categories*, *description*, etc. The underlying assumption is that this information may be useful in distinguishing laureates from non-laureates.

## Setting up the Environment

An initialization step is needed to setup the environment:
- The locale needs to be set for all categories to the user’s default setting (typically specified in the LANG environment variable) to enable correct sorting of words with accents.

import locale

locale.setlocale(locale.LC_ALL, '')

import copy
from datetime import datetime

import pandas as pd
from pandas.io.json import json_normalize

from src.data.dbpedia_utils import construct_resource_urls
from src.data.dbpedia_utils import find_resource_url
from src.data.dbpedia_utils import get_source_url
from src.data.dbpedia_utils import impute_redirect_filenames
from src.data.dbpedia_utils import json_categories_to_dict
from src.data.dbpedia_utils import json_keys_to_dict
from src.data.dbpedia_utils import json_values_to_dict
from src.data.dbpedia_utils import PHYSICISTS_IGNORE_REDIRECT_KEYS
from src.data.dbpedia_utils import PHYSICISTS_IMPUTE_KEYS
from src.data.jsonl_utils import read_jsonl
from src.data.progress_bar import progress_bar
from src.data.url_utils import get_filename_from_url
from src.data.url_utils import get_redirect_urls

## Reading in the JSON Lines Data

First let's read the JSON lines data into a list so that we can parse it later and take a look at the first entry.

json_lines = read_jsonl('../data/raw/physicists.jsonl.gz')
json_lines[0]

## Defining the Fields of Interest

Now we define the long list of keys and values that we wish to extract from the JSON lines data. Also, we define some that we explicitly wish to exclude. These keys and values are in the form of [semantic URLs](https://en.wikipedia.org/wiki/Clean_URL), which makes them easy to read and interpret. It also allows anyone to visit the resource in a web browser and see their meaning. The URLs are in 5 namespaces:

- [DBpedia Ontology](https://wiki.dbpedia.org/services-resources/ontology)
- DBpedia Property
- PURL
- W3
- FOAF

DBPEDIA_JSON_KEYS = [
    # DBpedia ontology
    'http://dbpedia.org/ontology/abstract',
    'http://dbpedia.org/ontology/academicAdvisor',
    'http://dbpedia.org/ontology/almaMater',
    'http://dbpedia.org/ontology/award',
    'http://dbpedia.org/ontology/birthDate',
    'http://dbpedia.org/ontology/birthName',
    'http://dbpedia.org/ontology/birthPlace',
    'http://dbpedia.org/ontology/child',
    'http://dbpedia.org/ontology/citizenship',
    'http://dbpedia.org/ontology/deathDate',
    'http://dbpedia.org/ontology/deathPlace',
    'http://dbpedia.org/ontology/doctoralAdvisor',
    'http://dbpedia.org/ontology/doctoralStudent',
    'http://dbpedia.org/ontology/education',
    'http://dbpedia.org/ontology/field',
    'http://dbpedia.org/ontology/influenced',
    'http://dbpedia.org/ontology/knownFor',
    'http://dbpedia.org/ontology/nationality',
    'http://dbpedia.org/ontology/notableStudent',
    'http://dbpedia.org/ontology/parent',
    'http://dbpedia.org/ontology/residence',
    'http://dbpedia.org/ontology/spouse',
    'http://dbpedia.org/ontology/thumbnail',
    'http://dbpedia.org/ontology/wikiPageID',
    'http://dbpedia.org/ontology/wikiPageRevisionID',

    # DBpedia property
    'http://dbpedia.org/property/academicAdvisor',
    'http://dbpedia.org/property/almaMater',
    'http://dbpedia.org/property/award',
    'http://dbpedia.org/property/birthDate',
    'http://dbpedia.org/property/birthName',
    'http://dbpedia.org/property/birthPlace',
    'http://dbpedia.org/property/child',
    'http://dbpedia.org/property/children',
    'http://dbpedia.org/property/citizenship',
    'http://dbpedia.org/property/deathDate',
    'http://dbpedia.org/property/deathPlace',
    'http://dbpedia.org/property/doctoralAdvisor',
    'http://dbpedia.org/property/doctoralStudent',
    'http://dbpedia.org/property/education',
    'http://dbpedia.org/property/field',
    'http://dbpedia.org/property/influenced',
    'http://dbpedia.org/property/knownFor',
    'http://dbpedia.org/property/nationality',
    'http://dbpedia.org/property/notableStudent',
    'http://dbpedia.org/property/parent',
    'http://dbpedia.org/property/residence',
    'http://dbpedia.org/property/signature',
    'http://dbpedia.org/property/spouse',
    'http://dbpedia.org/property/thesisTitle',
    'http://dbpedia.org/property/thesisUrl',
    'http://dbpedia.org/property/thesisYear',
    'http://dbpedia.org/property/thumbnail',
    'http://dbpedia.org/property/workInstitutions',
    'http://dbpedia.org/property/workplaces',

    # PURL
    'http://purl.org/dc/terms/description',

    # W3
    'http://www.w3.org/2000/01/rdf-schema#comment',
    'http://www.w3.org/ns/prov#wasDerivedFrom',

    # FOAF
    'http://xmlns.com/foaf/0.1/gender',
    'http://xmlns.com/foaf/0.1/givenName',
    'http://xmlns.com/foaf/0.1/isPrimaryTopicOf',
    'http://xmlns.com/foaf/0.1/name',
    'http://xmlns.com/foaf/0.1/surname'
]

DBPEDIA_JSON_VALUES = [
    # DBpedia ontology
    'http://dbpedia.org/ontology/influenced',  # target is influencedBy
    'http://dbpedia.org/ontology/influencedBy',  # target is influenced

    # DBpedia property
    'http://dbpedia.org/property/theorized'
]

DBPEDIA_IGNORE_URLS = [
    'http://dbpedia.org/resource/Bachelor_of_Arts',
    'http://dbpedia.org/resource/Bachelor_of_Science',
    'http://dbpedia.org/resource/Doctor_of_Philosophy',
    'http://dbpedia.org/resource/Master_of_Arts',
    'http://dbpedia.org/resource/Master_of_Philosophy',
    'http://dbpedia.org/resource/Master_of_Science',
    'http://dbpedia.org/resource/None',
    'http://dbpedia.org/resource/PhD',
    'http://dbpedia.org/resource/Ph.D',
    'http://dbpedia.org/resource/Ph.D.'
]

As you can see, there are several duplicate keys in the ontology and property namespaces (e.g. *citizenship*). Also, there are a lot of slightly different named keys that sound like they hold similar information (e.g. *alma mater* and *education*). Clearly, a set of rules will need to be defined about how to handle these duplicate keys and name variants.

## Creating the Physicists Dictionaries

Now we will parse the JSON lines data to create dictionaries. The following rules apply when creating the dictionaries:

1. Values in the DBpedia Ontology namespace takes precedence over those in the DBpedia Property namespace since, as described in section 4.3 of [DBpedia datasets](https://wiki.dbpedia.org/services-resources/datasets/dbpedia-datasets#h434-10), it contains the cleanest data. A Property namespace value is only used when there is no corresponding ontology namespace value.

2. Several keys holding similar data are merged together. *Education* is merged into *alma mater*, *work institutions* is merged into *workplaces* and *children* is merged into *child*. *InfluencedBy* and *Influenced* are also merged appropriately.

3. Some of the fields (e.g. the *abstract*) is multilingual. In such cases, only the English is extracted.

4. The data is messy and some cleanup is done on the fields involving *dates*, *nationality*, *citizenship*, *spouse* and *children*. However, a lot of noise remains in the data. The sources of the noise are: 
    - Fields containing semantic URLs that are redirected.
    - Semi-structured text which contains no information content (e.g. *stopwords*).
    - Semi-structured text containing valuable information that is not easy for a machine to understand. Either the text mixes concepts which should be in multiple fields or contains variant representations that essentially refer to the same "thing".
    
We will handle some of these issues now and some of them later, prior to and when generating features for machine learning. OK let's now generate the dictionaries and take a look at the first entry.

def create_physicist_data(json_line):
    """Create physicist data from json_line data.

    Args:
        json_line (dict): JSON dict.
    Returns:
        dict: Dictionary of physicist data.

    """

    # flatten the json
    flat_json = json_normalize(json_line)

    # find the resource, source and fullName
    resource_url = find_resource_url(flat_json)
    source_url = get_source_url(resource_url)
    full_name = get_filename_from_url(resource_url).replace('_', ' ')

    # construct the dictionary
    dict_ = {'resource': resource_url, 'source': source_url, 'fullName': full_name,
             **json_keys_to_dict(resource_url, flat_json,
                                 DBPEDIA_JSON_KEYS,
                                 ignore_urls=DBPEDIA_IGNORE_URLS),
             **json_values_to_dict(resource_url, flat_json, DBPEDIA_JSON_VALUES),
             **json_categories_to_dict(flat_json)}

    # merge keys
    dict_ = _merge_keys(dict_)

    # clean fields
    dict_ = _clean_fields(dict_)

    return dict_


def _merge_keys(dict_):
    merged_dict = copy.deepcopy(dict_)

    merged_dict = _merge_influences(merged_dict)
    merged_dict = _merge_right_key_into_left_key(
        merged_dict, 'almaMater', 'education')
    merged_dict = _merge_right_key_into_left_key(
        merged_dict, 'workplaces', 'workInstitutions')
    merged_dict = _merge_right_key_into_left_key(
        merged_dict, 'child', 'children')
    return merged_dict


def _clean_fields(dict_):
    cleaned_dict = copy.deepcopy(dict_)

    cleaned_dict = _clean_dates(cleaned_dict, 'birthDate')
    cleaned_dict = _clean_dates(cleaned_dict, 'deathDate')
    cleaned_dict = _clean_citizenship_nationality(cleaned_dict,
                                                  'citizenship')
    cleaned_dict = _clean_citizenship_nationality(cleaned_dict,
                                                  'nationality')
    cleaned_dict = _clean_spouse_child(cleaned_dict, 'spouse')
    cleaned_dict = _clean_spouse_child(cleaned_dict, 'child')
    cleaned_dict = _clean_almaMater(cleaned_dict)
    return cleaned_dict


def _merge_influences(dict_):
    merged_dict = copy.deepcopy(dict_)

    influencedBy = merged_dict.get('influencedBy')
    influenced = merged_dict.get('influenced')
    influenced_ = merged_dict.get('influenced_')

    if influenced_:
        if influencedBy:
            merged_dict['influencedBy'] += '|' + influenced_
        else:
            merged_dict['influencedBy'] = influenced_
        del merged_dict['influenced_']

    influencedBy = merged_dict.get('influencedBy')
    if influencedBy and influenced:
        merged_dict['influencedBy'], merged_dict['influenced'] = \
            merged_dict['influenced'], merged_dict['influencedBy']
    elif influencedBy and not influenced:
        merged_dict['influenced'] = influencedBy
        del merged_dict['influencedBy']
    elif not influencedBy and influenced:
        merged_dict['influencedBy'] = influenced
        del merged_dict['influenced']
    
    influencedBy = merged_dict.get('influencedBy')
    influenced = merged_dict.get('influenced')
    if influencedBy:
        merged_dict['influencedBy'] = '|'.join(sorted(
            merged_dict['influencedBy'].split('|'), key=locale.strxfrm))
    if influenced:
        merged_dict['influenced'] = '|'.join(sorted(
            merged_dict['influenced'].split('|'), key=locale.strxfrm))

    return merged_dict


def _merge_right_key_into_left_key(dict_, left_key, right_key):
    merged_dict = copy.deepcopy(dict_)

    left = merged_dict.get(left_key)
    right = merged_dict.get(right_key)

    if not left and right:
        merged_dict[left_key] = right
    merged_dict.pop(right_key, None)
    return merged_dict


def _clean_citizenship_nationality(dict_, key):
    cleaned_dict = copy.deepcopy(dict_)

    key_present = cleaned_dict.get(key)
    if (key_present and not
        key_present == 'United Kingdom of Great Britain and Ireland'):
        cleaned_dict[key] = (
            key_present
            .replace(' / ', '|')
            .replace(' , ', '|')
            .replace(', ', '|')
            .replace(' and then ', '|')
            .replace(' and ', '|')
            .replace('Citizenship in the ', '')
            .replace('Citizen of ', '')
            .replace(' citizen', '')
            .replace(' law', '')
            .replace('List of ', '')
            .replace('Nationality law of the ', '')
            .replace('Nationality of the ', '')
            .replace(' nationality law', '')
            .replace(' nationality', '')
            .replace('People of ', '')
            .replace(' people', '')
            .replace(' subject', '')
        )
        cleaned_dict[key] = '|'.join(sorted(cleaned_dict[key].split('|'),
                                            key=locale.strxfrm))
    return cleaned_dict


def _clean_almaMater(dict_):
    cleaned_dict = copy.deepcopy(dict_)

    key_present = cleaned_dict.get('almaMater')
    if key_present:
        cleaned_dict['almaMater'] = (
             key_present
            .replace('|Theoretical Physics', '')
            .replace('|Physics', '')
        )
    return cleaned_dict


def _clean_spouse_child(dict_, key):
    cleaned_dict = copy.deepcopy(dict_)

    key_present = cleaned_dict.get(key)
    if key_present and isinstance(key_present, str):
        cleaned_dict[key] = (
            key_present
            .replace(', and ', '|')
            .replace(' and ', '|')
            .replace(', ', '|')
            .replace(' ; 1 child', '')
            .replace('|divorced', '')
            .replace('divorced|', '')
            .replace('sons ', '')
        )
        cleaned_dict[key] = '|'.join(sorted(cleaned_dict[key].split('|'),
                                            key=locale.strxfrm))
    return cleaned_dict


def _clean_dates(dict_, key):
    cleaned_dict = copy.deepcopy(dict_)

    date = dict_.get(key)
    if date:
        # replace invalid months and days with first day and first month
        # the dates only need to be approximately correct as years lived
        # will be estimated from them
        if isinstance(date, int):  # year only
            first_date = str(date) + '-01-01'
        else:
            first_date = date.split('|')[0]
            first_date = first_date.replace('-0-0', '-01-01')
            if first_date.endswith('-0'):
                first_date = first_date[:-2] + '-1'
            if 'c. ' in date:  # deal with circa
                first_date = date.replace('c. ', '') + '-01-01'
            # ignore dates with a year below 1000 as datetime does not
            # handle these
            if (first_date.startswith('-') or
                int(first_date.split('-')[0]) < 1000):
                return cleaned_dict
        cleaned_dict[key] = str(
            datetime.strptime(first_date, '%Y-%m-%d').date())
    return cleaned_dict

data = [create_physicist_data(json_line) for json_line in json_lines]
data[0]

You can see that many of the values in the dictionary contain lists of semantic URLs that are each meant to refer to a unique "thing" such as a person or a place. If the URL is redirected, it is important to know where it is redirected to, so that identical "things" resolve to the same URL.

## Imputing Redirects in the Physicists Dictionaries

In order to impute redirects in the dictionaries, we need to perform the following steps:

1. **Parse the dictionaries to obtain a list of all URLs**. These are either resource URLs or URLs constructed in an *ad hoc* manner from the free form text in the hope that they resolve to a genuine resource URL. The aim is to resolve as many items as possible in the fields to semantic URLs. We restrict URL selection and construction to the following impute keys: `academicAdvisor`, `almaMater`, `award`, `birthPlace`, `categories`, `child`, `citizenship`, `deathPlace`, `doctoralAdvisor`, `doctoralStudent`, `field`, `influenced`, `influencedBy`, `knownFor`, `nationality`, `notableStudent`, `parent`, `residence`, `spouse`, `theorized` and `workplaces` as these are fields of interest involving URLs that features may be extracted from.
2. **Submit HTTP requests to fetch the URLs and determine the redirects**. A cache is kept mapping the URLs to their redirects. As a consequence a HTTP request is only made to fetch a URL if it is not found in the cache. This greatly helps with performance, as does the fact that the code is asynchronous.
3. **Replace the URLs in the dictionaries with their redirects**. In fact we use just the filename since the paths are identical for every URL.

urls_to_check = construct_resource_urls(data, PHYSICISTS_IMPUTE_KEYS)
len(urls_to_check)

url_cache_path = '../data/raw/dbpedia-redirects.csv'
redirects = get_redirect_urls(
    urls_to_check, url_cache_path=url_cache_path, max_workers=20, timeout=30,
    progress_bar=progress_bar(len(urls_to_check)))
len(redirects)

You can see that many of the requested URLs were not found. This is to be expected as there is free form text in the fields that do not map to semantic URLs in DBpedia. However, this *ad hoc* approach of constructing URLs from free form texts does find legitimate URLs in many instances. Now let's sort and persist the URL cache to disk in case any new URLs are found.

dbpedia_redirects = pd.DataFrame(
    sorted(list(zip(redirects.keys(), redirects.values())),
           key=lambda x: locale.strxfrm(x[0])), columns=['url', 'redirect_url'])
dbpedia_redirects.to_csv(url_cache_path, index=False)

Now we will replace the URLs in the dictionaries with their redirects making sure to just use the filename as the paths are identical for every URL.

imputed_data = impute_redirect_filenames(data, PHYSICISTS_IMPUTE_KEYS, redirects,
                                         PHYSICISTS_IGNORE_REDIRECT_KEYS)
imputed_data[0]

## Creating the Physicists Dataframe

Now we use the dictionaries of imputed data to create a dataframe. Let's confirm that it contains the expected number of physicists and take a look at it.

physicists = pd.DataFrame(imputed_data)
assert(len(physicists) == 1058)
with pd.option_context('display.max_rows', 1100):
    display(physicists)

## Persisting the Data

Now we have the dataframe, let's persist it for later analysis by writing its contents to a csv file.

physicists.to_csv('../data/interim/physicists.csv', index=False)
